refactor: 컨트롤러 구조 변경
Rest api와 페이지 렌더링 컨트롤러 분리
